+++ metadata.rb
@@ -4,19 +4,21 @@ license 'Apache 2.0'
 description 'installs/configures dop_base'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 name 'dop_base'
-version '0.2.14'
+version '0.4.0'
+issues_url 'https://github.com/ffuenf/dop_base/issues'
+source_url 'https://github.com/ffuenf/dop_base'
 
-%w( debian ).each do |os|
+%w( debian ubuntu ).each do |os|
   supports os
 end
 
 %w(
+  build-essential
   sudo
   user
   sshd
   ssh
   hostnames
-  hostsfile
   apt
   git
   dotdeb